chatbot-app-docker/backend/requirements.txt
annotated-types==0.7.0
anyio==4.12.1
certifi==2026.2.25
click==8.1.8
distro==1.9.0
dotenv==0.9.9
exceptiongroup==1.3.1
fastapi==0.115.0
gunicorn==23.0.0
h11==0.16.0
httpcore==1.0.9
httptools==0.7.1
httpx==0.28.1
idna==3.11
jiter==0.13.0
openai==2.29.0
packaging==26.0
psycopg2-binary==2.9.9
pydantic==2.12.5
pydantic_core==2.41.5
python-dotenv==1.0.1
PyYAML==6.0.3
sniffio==1.3.1
starlette==0.38.6
tqdm==4.67.3
typing-inspection==0.4.2
typing_extensions==4.15.0
uvicorn==0.30.6
uvloop==0.22.1
watchfiles==1.1.1
websockets==15.0.1
prometheus-client==0.24.1
